requirements.txt
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cupshelpers==1.0
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro-info===0.18ubuntu0.18.04.1
Django==3.2.18
django-allauth==0.52.0
django-ckeditor==6.3.2
django-debug-toolbar==3.2.4
django-filter==21.1
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
httplib2==0.9.2
idna==2.6
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.4.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Markdown==3.3.7
MarkupSafe==2.0.1
mysqlclient==2.1.1
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.2.2
olefile==0.45.1
packaging==21.3
pexpect==4.2.1
Pillow==8.4.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.9
pyRFC3339==1.0
python-apt==1.6.6
python-dateutil==2.6.1
python-debian==0.1.32
python3-openid==3.2.0
pytz==2022.7.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
system-service==0.3
systemd-python==234
typing_extensions==4.1.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.22
virtualenv==15.1.0
wadllib==1.3.2
whitenoise==5.3.0
wrapt==1.14.1
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
